+++ src/services/AccountManager.cpp
@@ -1,5 +1,6 @@
 #include "services/AccountManager.hpp"
 
+#include <spdlog/logger.h>
 #include <boost/uuid/random_generator.hpp>
 #include <boost/uuid/uuid.hpp>
 #include <boost/uuid/uuid_io.hpp>